Move FakeChatService into its own file
